Revert "complete auth routers"
