+++ auth_service/app/db.py
@@ -1,3 +1,5 @@
+import os
+
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 
@@ -5,4 +7,5 @@ db = SQLAlchemy()
 
 
 def init_db(app: Flask):
+    app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('POSTGRES_URL')
     db.init_app(app)
